prepare for CPAN
I'll have to contact the kwalitee people to get some last issues resolved

Here is a list of failed Kwalitee tests and
what you can do to solve them:

* has_license_in_source_file
Add =head1 LICENSE and the text of the license to the main module in your code.

* use_strict
Add 'use strict' (or its equivalents) to all modules, or convince us that your favorite module is well-known enough and people can easily see the modules are strictly written.


Failed optional Kwalitee tests and
what you can do to solve them:

* has_known_license_in_source_file
Add =head1 LICENSE and/or the proper text of the well-known license to the main module in your code.